% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz02Functions.R
\name{accuracy}
\alias{accuracy}
\title{Calculate accuracy}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{number or boolean vector of predicted values (0 or F = negative, 1 or T = positive)}

\item{truth}{number or boolean vector of true values (0 or F = negative, 1 or T = positive)}
}
\value{
The accuracy, formula: accuracy = (# True Positives + # True Negatives) / (# All Positives + # All Negatives)
}
\description{
Calculate accuracy
}
\examples{
accuracy(c(0,1,0),c(1,0,1))
accuracy(c(1,1,1),c(1,0,0))
}
