% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw11Functions.R
\name{adjR2}
\alias{adjR2}
\title{Calculate Adjusted R-Squared}
\usage{
adjR2(pred, truth, d)
}
\arguments{
\item{pred}{numerical vector of predicted values}

\item{truth}{numerical vector of true values}

\item{d}{number of predictor variables used in model}
}
\value{
Adjusted r-squared value as a measure of model fitness. formula: Adjusted R^2 = 1 - ((RSS/(n-d-1))/(TSS/(n-1)))
}
\description{
Calculate Adjusted R-Squared
}
\examples{
yTrue = c(0,1,1,1)

adjR2(pred = c(0,1,1,1.1), truth = yTrue, d = 1)
adjR2(pred = c(0,1,0.7,1), truth = yTrue, d = 2)
}
