% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04Functions.R
\name{boxMuller}
\alias{boxMuller}
\title{Calculate Box-Muller Transformation}
\usage{
boxMuller(n)
}
\arguments{
\item{n}{number of samples to take from the standard normal distribution}
}
\value{
2 vectors in matrix columns 1 and 2. x1 and x2 are independent, normally distributed random vectors with n values each
}
\description{
Calculate Box-Muller Transformation
}
\examples{
boxMuller(5)
boxMuller(100)
}
