% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw07Functions.R
\name{chiSquareCounts}
\alias{chiSquareCounts}
\title{Perform ChiSquare Test of Homogeneity}
\usage{
chiSquareCounts(tib, x, y)
}
\arguments{
\item{tib}{tibble dataframe with colnames that contain x and y}

\item{x}{column x of tib, variable 1 of comparison}

\item{y}{column y of tib, variable 2 of comparison}
}
\value{
p-value from ChiSquare Test of Homogenity between data in columns x and y of tib
}
\description{
Perform ChiSquare Test of Homogeneity
}
\examples{
data = tibble::tibble("sex" = c(1,0,1),"group" = c("A","B","C"), "age" = c(40,15,33), "height" = c(63, 70, 68))

chiSquareCounts(data, "sex", "group")
chiSquareCounts(data, "age", "height")
}
