% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw06Functions.R
\name{effectSize}
\alias{effectSize}
\title{Calculate Effect Size}
\usage{
effectSize(x, g)
}
\arguments{
\item{x}{vector of values of length n}

\item{g}{vector of group designations (0 or 1) of length n, corresponding to vector x}
}
\value{
effect size, or Cohen's d. formula: d = |mean0 - mean1| / sqrt((sd0^2 + sd1^2)/2)
}
\description{
Calculate Effect Size
}
\examples{
effectSize(c(1.42, 8.10, 9.22), c(0,1,1))
effectSize(c(1.42, 8.10, 9.22, 20.92, 5.67, 0.45), c(0,0,1,1,0))
}
