% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz02Functions.R
\name{f1}
\alias{f1}
\title{Calculate F1 Score}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{number or boolean vector of predicted values (0 or F = negative, 1 or T = positive)}

\item{truth}{number or boolean vector of true values (0 or F = negative, 1 or T = positive)}
}
\value{
The F1 Score, formula: F1 = 2 * (precision * recall) / (precision + recall)
}
\description{
Calculate F1 Score
}
\examples{
f1(c(0,1,0),c(1,0,1))
f1(c(1,1,1),c(1,0,0))
}
