% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw07Functions.R
\name{minimumN}
\alias{minimumN}
\title{Calculate minimum number of samples}
\usage{
minimumN(d)
}
\arguments{
\item{d}{target effect size}
}
\value{
minimum number of samples needed to achieve effect size d and power = 0.8
}
\description{
Calculate minimum number of samples
}
\examples{
minimumN(1.3)
minimumN(0.5)
}
