% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{pcApprox}
\alias{pcApprox}
\title{Approximate Data from N Principle Components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{numerical dataframe, matrix, or tibble}

\item{npc}{number of Principle Components (PCs) to use in approximation}
}
\value{
approximation of data x based on npc principal components
}
\description{
Approximate Data from N Principle Components
}
\examples{
x = matrix(c(4,5,2,10,2,8,4,5,6), ncol = 3)
pcApprox(x, npc = 1)
}
