% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{pcLollipop}
\alias{pcLollipop}
\title{Create Lollipop Plots of Principal Component Loadings}
\usage{
pcLollipop(x)
}
\arguments{
\item{x}{numerical dataframe, or tibble}
}
\value{
plotting object for lollipop plots of principal components of x
}
\description{
Create Lollipop Plots of Principal Component Loadings
}
\examples{
x = data.frame(data = matrix(c(4,5,2,10,2,8,4,5,6), ncol = 3))
plot = pcLollipop(x)
}
