% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw07Functions.R
\name{postHocPower}
\alias{postHocPower}
\title{Estimate of Post Hoc Power from 1000 Simulations}
\usage{
postHocPower(d, n1, n2)
}
\arguments{
\item{d}{effect size (Cohen's d)}

\item{n1}{number of samples in group 1}

\item{n2}{number of samples in group 2}
}
\value{
estimate of post hoc power from 1000 simulations. formula: power estimate = # statistically significant simulations / total # of simulations
}
\description{
Estimate of Post Hoc Power from 1000 Simulations
}
\examples{
postHocPower(d = 1, n1 = 100, n2 = 90)
postHocPower(d = 0.5, n1 = 20, n2 = 22)
}
