% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz02Functions.R
\name{ppv}
\alias{ppv}
\title{Calculate PPV}
\usage{
ppv(pred, truth)
}
\arguments{
\item{pred}{number or boolean vector of predicted values (0 or F = negative, 1 or T = positive)}

\item{truth}{number or boolean vector of true values (0 or F = negative, 1 or T = positive)}
}
\value{
The PPV, formula: PPV = # True Positives / (# True Positives + # False Positives)
}
\description{
Calculate PPV
}
\examples{
ppv(c(0,1,0),c(1,0,1))
ppv(c(1,1,1),c(1,0,0))
}
