% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw11Functions.R
\name{r2}
\alias{r2}
\title{Calculate R-Squared}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{numerical vector of predicted values}

\item{truth}{numerical vector of true values}
}
\value{
R-squared value as a measure of model fitness. formula: R^2 = 1 - RSS/TSS
}
\description{
Calculate R-Squared
}
\examples{
yTrue = c(0,1,1,1)

r2(pred = c(0,1,1,1.1), truth = yTrue)
r2(pred = c(0,1,0.7,1), truth = yTrue)
}
