% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz02Functions.R
\name{specificity}
\alias{specificity}
\title{Calculate specificity}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{number or boolean vector of predicted values (0 or F = negative, 1 or T = positive)}

\item{truth}{number or boolean vector of true values (0 or F = negative, 1 or T = positive)}
}
\value{
The specificity, formula: specificity = # True Negatives / (# True Negatives + # False Positives)
}
\description{
Calculate specificity
}
\examples{
specificity(c(0,1,0),c(1,0,1))
specificity(c(1,1,1),c(1,0,0))
}
