% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw06Functions.R
\name{twoSidedT}
\alias{twoSidedT}
\title{Calculate p-value from Two-Sided T-Test}
\usage{
twoSidedT(t, n)
}
\arguments{
\item{t}{t-test statistic}

\item{n}{sample size}
}
\value{
p-value, or area under the t-density calculated from t statistic and sample size
}
\description{
Calculate p-value from Two-Sided T-Test
}
\examples{
twoSidedT(3, 5)
twoSidedT(0.2, 2)
}
