% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw06Functions.R
\name{twoSidedZ}
\alias{twoSidedZ}
\title{Calculate p-value from Z-Score}
\usage{
twoSidedZ(z)
}
\arguments{
\item{z}{z-score}
}
\value{
p-value, or area under the z-density calculated from z score
}
\description{
Calculate p-value from Z-Score
}
\examples{
twoSidedZ(3)
twoSidedZ(0.2)
}
