% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{unscale}
\alias{unscale}
\title{Unscale}
\usage{
unscale(x)
}
\arguments{
\item{x}{numerical object from scale() with or without attributes}
}
\value{
un-scaled version of x. reversed centering/scaling of x, if any
}
\description{
Unscale
}
\examples{
xS = scale(matrix(1:10, ncol = 2))
unscale(x = xS)
}
