% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw06Functions.R
\name{welchT}
\alias{welchT}
\title{Perform Welch's T-Test}
\usage{
welchT(x, y)
}
\arguments{
\item{x}{vector of values of length n}

\item{y}{vector of values of length n}
}
\value{
returns vector of length 5 with Welch's T statistic, degrees of freedom, p-value, mean of group x, and mean of group y
}
\description{
Perform Welch's T-Test
}
\examples{
welchT(c(1.42, 8.10, 9.22), c(10.4, 4.52, 3.44))
welchT(c(1.42, 8.10, 9.22, 20.92, 5.67, 0.45), c(1.99, 8.09, 9.13, 9.10, 5.55, 0.11))
}
